requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
jupyter>=1.0.0

# Dependencias opcionales para GPU
cupy-cuda11x>=11.0.0  # Para GPUs NVIDIA con CUDA 11.x
# cupy-cuda12x>=12.0.0  # Para GPUs NVIDIA con CUDA 12.x (alternativa)

# Para instalar CuPy, descomenta la línea apropiada según tu versión de CUDA:
# pip install cupy-cuda11x  (para CUDA 11.x)
# pip install cupy-cuda12x  (para CUDA 12.x)